requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyMuPDF==1.23.8
reportlab==4.0.7
python-docx==1.1.0
Pillow==10.1.0
